Make controllers dir structure hierarchical
E.g. from "pkg/ipamctlr/ipamctlr.go" to "pkg/controllers/ipam/ipam.go".
Similar changes are made to all the directories where files of the
controllers (ipam and ca) reside or will reside.
